Toggle switch built for RSVP step
